Mutual Exclusion using Semaphore.c
Semaphores:
sem_init(&mutex, 0, 1);: Initialize a semaphore with a value of 1.
sem_wait(&mutex);: Decrement the semaphore. If the semaphore value is zero, the calling thread will block until the semaphore value is greater than zero.
sem_post(&mutex);: Increment the semaphore and wake up a blocked thread if any.